Made database go beep boop

+++ java/src/main/java/com/techelevator/dao/FriendDAO.java
@@ -9,6 +9,7 @@ import org.springframework.jdbc.support.rowset.SqlRowSet;
 import org.springframework.stereotype.Service;
 
 import com.techelevator.model.FriendDTO;
+import com.techelevator.model.User;
 
 @Service
 public class FriendDAO {
@@ -30,6 +31,17 @@ public class FriendDAO {
 
 		return friendDTO;
 	}
+	public List<User> getAllFriendsById(long id) {
+
+		List<User> friends = new ArrayList<>();
+		String sqlFindFriendID = "SELECT u.* FROM friend f INNER JOIN users u ON f.friend_id = u.user_id WHERE f.user_id = ?";
+		SqlRowSet friendRow = jdbcTemplate.queryForRowSet(sqlFindFriendID,id);
+		while (friendRow.next()) {
+			friends.add(mapRowToUser(friendRow));
+		}
+
+		return friends;
+	}
 	public boolean addFriend(FriendDTO friend, int userID) {
 		
 		String sqlInsertFriend = "INSERT INTO friend (user_id, friend_id, status_id, status_id_desc) " +
@@ -48,9 +60,9 @@ public class FriendDAO {
 	}
 	
 	public boolean deleteFriend(int friendID, int userID) {
-		String deleteFriend = "DELETE FROM friend WHERE friend_id =?";
+		String deleteFriend = "DELETE FROM friend WHERE friend_id =? AND user_id = ?";
 		
-		int success = jdbcTemplate.update(deleteFriend, friendID);
+		int success = jdbcTemplate.update(deleteFriend, friendID,userID);
 		return (success>0)? true:false;
 	}
 		
@@ -62,7 +74,12 @@ public class FriendDAO {
 		friend.setStatusIDDesc(friendRow.getString("status_id_desc"));
 		return friend;
 	}
-	
+	 private User mapRowToUser(SqlRowSet rs) {
+	        User user = new User();
+	        user.setId(rs.getLong("user_id"));
+	        user.setUsername(rs.getString("username"));
+	        return user;
+	    }
 }
